(0.2.5) Fix issue where the logger stops delivering messages to sumologic
The problem was triggered when the client would take too long to return from
log_event/2. Agent.update/2 has a default timeout of 5 seconds, and if the
provided function (which calls log_event/2) does not return, the calling
process is exited.

The solution is to defer the log_event/2-call to a process which has no
other purpose than handling the logging.